Uploaded existing project
Uploaded the existing Fallout 3 Terminal Hacking project

This was intended to be a little throw-away utility, so it was on my
local file system not under version control; since it became more of
a project with additional features and functionality being added and
changed, it is now here.

A rough development timeline is as follows:

* Haskell command-line prototype, simply removing ineligible
candidate words at each step.  Not optimal.
* English-language GUI for the above, displaying remaining
candidates, along with input for the word tested and the
result given in-game.
* Added i18n support.  Translated (poorly?) into Japanese.
* Better (optimal?) algorithm implemented, allows guessing
non-candidate words if it eliminates the greatest number
of remaining candidates.